+++ src/schemas/analytics.ts
