+++ src/schemas/memory.ts
