+++ src/schemas/user.ts
@@ -1,0 +1,16 @@
+import { z } from 'zod';
+import { apiKeySchema } from './apiKey.js';
+import { subscriptionSchema } from './subscriptions.js';
+
+
+export const userSchema = z.object({
+    id: z.string().uuid(),
+    username: z.string().min(3).max(30),
+    email: z.string().email(),
+    password: z.string().min(8).max(100),
+    createdAt: z.coerce.date(),
+    updatedAt: z.coerce.date().optional(),
+    apiKeys: z.array(apiKeySchema).optional(), // Optional array of API keys associated with the user
+    subscriptionPlans: z.array(subscriptionSchema).optional(), // Optional array of subscriptions associated with the user
+})
+export type User = z.infer<typeof userSchema>;
