+++ src/schemas/vector.ts
